Trying out new LSP structure
