+++ nvim/lua/plugins/lsp/configs/eslint.lua
@@ -1,0 +1,11 @@
+local M = {}
+
+M.setup = function(capabilities, on_attach)
+  return {
+    capabilities = capabilities,
+    on_attach = on_attach,
+    settings = {format = false}
+  }
+end
+
+return M
